feat: add /health endpoint, postgres db connection
Base structure is redefined, with the src/api/ being the folder with
the API logic, configs/ containing essential app, db and logging configs
A simple express application template to start the API server.
There are a few tweaks introduced to linting and Jest as well.
New dependencies are added to create a simple HTTP server with db conn.